requirements.txt
pandas==2.3.2
numpy
faiss-cpu==1.12.0
sentence-transformers==5.1.1
google-genai
